Update datepicker docs to clarify separate install of providers
The current document does not clearly convey that the date implementations are a separate install to the material framework.

This update clarifies that the installation of the modules is a separate step and provides additional context by providing an example command of a necessary step. This highlights the need to install the provider while retaining the module path information.